Refactoring, add the ability for an admin to view the group he belongs to
