+++ client/src/app/main-page/main-page.component.spec.ts
@@ -1,6 +1,6 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { MainPageViewComponent } from './main-page.component';
+import { MainPageViewComponent } from "../main-page-view/main-page-view.component";
 
 describe('MainPageViewComponent', () => {
   let component: MainPageViewComponent;
